Created the LoginViewModel and changed the location of the post for the User
