+++ RemoteMC3/LoginScreen/Views/LoginViewController.swift
@@ -9,11 +9,9 @@
 import UIKit
 import AuthenticationServices
 
-protocol LoginViewControllerDelegate {
-    func didFinishAuth()
-}
-
 class LoginViewController: UIViewController {
+    
+    var viewModel: LoginViewModel = LoginViewModel()
     
     var delegate: LoginViewControllerDelegate?
 
@@ -28,53 +26,8 @@ extension LoginViewController: ASAuthorizationControllerDelegate {
     // Implementacao no Servidor -> POST
     private func registerNewAccount(credential: ASAuthorizationAppleIDCredential) {
         print("Registering new account with user: \(credential.user)")
-        guard let firstName = credential.fullName?.givenName else {
-            return
-        }
-        guard let lastName = credential.fullName?.familyName else {
-            return
-        }
-        guard let email = credential.email else {
-            return
-        }
         
-        let fullName = firstName + " " + lastName
-        
-        let userInfo = ["name": "\(fullName)", "email": "\(email)"]
-        
-        if let url = URL(string: "https://projeta-server.herokuapp.com/createUser") {
-            let session = URLSession.shared
-            var request = URLRequest(url: url)
-            request.httpMethod = "POST"
-            
-            do {
-                request.httpBody = try JSONSerialization.data(withJSONObject: userInfo, options: .prettyPrinted)
-            } catch let error {
-                print(error.localizedDescription)
-            }
-            
-            request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-            request.addValue("application/json", forHTTPHeaderField: "Accept")
-            
-            session.dataTask(with: request as URLRequest, completionHandler: { data, response, error in
-                guard error == nil else {
-                    return
-                }
-
-                guard let data = data else {
-                    return
-                }
-                
-                do {
-                    //create json object from data
-                    if let json = try JSONSerialization.jsonObject(with: data, options: .mutableContainers) as? [String: Any] {
-                        print(json)
-                    }
-                } catch let error {
-                    print(error.localizedDescription)
-                }
-                }).resume()
-        }
+        viewModel.saveUser(credential: credential)
         
         delegate?.didFinishAuth()
         self.dismiss(animated: true, completion: nil)
@@ -172,7 +125,6 @@ extension LoginViewController: ASAuthorizationControllerDelegate {
             backgroundView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
             backgroundView.heightAnchor.constraint(equalTo: view.safeAreaLayoutGuide.heightAnchor, multiplier: 0.45),
 			iconImage.topAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.topAnchor, constant: 20),
-//            iconImage.centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: -150),
             iconImage.centerXAnchor.constraint(equalTo: view.centerXAnchor),
 			iconImage.heightAnchor.constraint(equalToConstant: 300),
             iconImage.widthAnchor.constraint(equalToConstant: 300),
